# server/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/tenants": {
      "post": {
        "operationId": "TenantsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTenantDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateTenantResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "テナント作成（初期管理者設定含む）",
        "tags": [
          "tenants"
        ]
      },
      "get": {
        "operationId": "TenantsController_findAll",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantListResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "テナント一覧",
        "tags": [
          "tenants"
        ]
      }
    },
    "/api/v1/tenants/{id}": {
      "get": {
        "operationId": "TenantsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantWithCountResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "テナント詳細",
        "tags": [
          "tenants"
        ]
      },
      "patch": {
        "operationId": "TenantsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTenantDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "テナント更新",
        "tags": [
          "tenants"
        ]
      }
    },
    "/api/v1/tenants/{id}/suspend": {
      "post": {
        "operationId": "TenantsController_suspend",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "テナント停止",
        "tags": [
          "tenants"
        ]
      }
    },
    "/api/v1/tenants/{id}/resume": {
      "post": {
        "operationId": "TenantsController_resume",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "テナント再開",
        "tags": [
          "tenants"
        ]
      }
    },
...
    "/api/v1/admin/departments": {
      "post": {
        "operationId": "SettingsController_createDepartment",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDepartmentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DepartmentResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "部署作成",
        "tags": [
          "settings"
        ]
      },
      "get": {
        "operationId": "SettingsController_findDepartments",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DepartmentResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "部署一覧",
        "tags": [
          "settings"
        ]
      }
    },
    "/api/v1/admin/departments/{id}": {
      "patch": {
        "operationId": "SettingsController_updateDepartment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDepartmentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DepartmentResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "部署更新",
        "tags": [
          "settings"
        ]
      }
    },
    "/api/v1/admin/work-patterns": {
      "post": {
        "operationId": "SettingsController_createWorkPattern",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkPatternDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkPatternResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "勤務形態作成",
        "tags": [
          "settings"
        ]
      },
      "get": {
        "operationId": "SettingsController_findWorkPatterns",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WorkPatternResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "勤務形態一覧",
        "tags": [
          "settings"
        ]
      }
    },
    "/api/v1/admin/work-patterns/{id}": {
      "patch": {
        "operationId": "SettingsController_updateWorkPattern",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkPatternDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkPatternResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "勤務形態更新",
        "tags": [
          "settings"
        ]
      }
    }
  },
  "info": {
    "title": "勤怠管理システム API",
    "description": "マルチテナント勤怠管理システムのREST API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateTenantDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "株式会社サンプル"
          },
          "adminEmail": {
            "type": "string",
            "example": "admin@example.com"
          },
          "adminName": {
            "type": "string",
            "example": "田中太郎"
          }
        },
        "required": [
          "name",
          "adminEmail",
          "adminName"
        ]
      },
      "TenantResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "suspended"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "name",
          "status",
          "createdAt",
          "updatedAt"
        ]
      },
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "system_admin",
              "tenant_admin",
              "tenant_user"
            ]
          },
          "isActive": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "firebaseUid": {
            "type": "string"
          },
          "tenantId": {
            "type": "object",
            "nullable": true
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "email",
          "name",
          "role",
          "isActive",
          "createdAt",
          "firebaseUid",
          "tenantId",
          "updatedAt"
        ]
      },
      "CreateTenantResponseDto": {
        "type": "object",
        "properties": {
          "tenant": {
            "$ref": "#/components/schemas/TenantResponseDto"
          },
          "adminUser": {
            "$ref": "#/components/schemas/UserResponseDto"
          }
        },
        "required": [
          "tenant",
          "adminUser"
        ]
      },
      "TenantUserCountResponseDto": {
        "type": "object",
        "properties": {
          "users": {
            "type": "number"
          }
        },
        "required": [
          "users"
        ]
      },
      "TenantWithCountResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "suspended"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "_count": {
            "$ref": "#/components/schemas/TenantUserCountResponseDto"
          }
        },
        "required": [
          "id",
          "name",
          "status",
          "createdAt",
          "updatedAt",
          "_count"
        ]
      },
      "TenantListResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TenantWithCountResponseDto"
            }
          },
          "total": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          }
        },
        "required": [
          "data",
          "total",
          "page",
          "limit"
        ]
      },
      "UpdateTenantDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "株式会社サンプル更新"
          }
        }
      },
...
      "DepartmentResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "tenantId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "tenantId",
          "name"
        ]
      },
      "WorkPatternResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "tenantId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "startTime": {
            "type": "string"
          },
          "endTime": {
            "type": "string"
          },
          "breakMinutes": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "tenantId",
          "name",
          "startTime",
          "endTime",
          "breakMinutes"
        ]
      },
...
      "CreateDepartmentDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "開発部"
          }
        },
        "required": [
          "name"
        ]
      },
      "CreateWorkPatternDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "通常勤務"
          },
          "startTime": {
            "type": "string",
            "example": "09:00:00",
            "description": "所定開始時刻（HH:MM:SS）"
          },
          "endTime": {
            "type": "string",
            "example": "18:00:00",
            "description": "所定終了時刻（HH:MM:SS）"
          },
          "breakMinutes": {
            "type": "number",
            "example": 60,
            "description": "所定休憩時間（分）"
          }
        },
        "required": [
          "name",
          "startTime",
          "endTime"
        ]
      }
    }
  }
}